Refactoring stations dataframe cleaning clode
